requirements.txt
# 会议聊天室系统依赖
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
